Fix a bug where loading a project which contained a Knob with a single dimension would not load properly with another number of dimensions
There is a case where the dimension of a parameter might have changed
between versions, e.g:
the size parameter of the Blur node was previously a Double1D and has
become a Double2D to control
both dimensions.
For compatibility, we do not load only the first dimension, otherwise
the result wouldn't be the same,
 instead we replicate the last dimension of the serialized knob to all
other remaining dimensions to fit the
knob's dimensions.