RenderRoI: cache the hash with rounded time for non continuous effects. Also getImage:
If an image was not requested in getFramesNeeded:

If we are in the render action:
This node did not request anything on the input in getFramesNeeded
action: it did a raw call to fetchImage without advertising us.
Don't fail:
If we are in a current render, (tls->currentRenderArgs.validArgs==true)
 call getRegionsOfInterest on the current render window
If are not in a render, well fall back onto rendering the whole image
(RoD).
If we are in an analysis, the effect does not necessarily have to use
getFramesNeeded, but the ctor of ParallelRenderArgsSetter in this
function should have added a request
at the exact time passed in parameter of this function.