fix #279: fix ticks computation
In -Ofast compilation mode the computation of std::pow(10., x)
is not perfect and not gives an exact number due to floating
point number approximations.
This result in wrong behaviour later as in timeline ticks drawing.

As we only need to compute an integer computation, replace
the double 10. per the integer 10 to obtain a correct value.

Signed-off-by: Guillaume Roguez <yomgui1@gmail.com>