Gui/DockablePanel: connect to returnPressed instead of editingFinished for _nameLineEdit
QTBUG-40

QLineEdit triggers editingFinished when focus changes, this creates issues when we open a dialog based on this signal and return focus to the widget when done.

Example: The node rename error dialog is triggered twice and after that focus is returned to the QLineEdit, on Windows this will trigger editingFinished again, resulting in an endless loop.