create GuiApp.pasteNodes() python function
This function paste the nodes in the clipboard.
It will be used to reconnect the pasted nodes to the previously selected node as Nuke does. (good time saver)

Known limitation: the nodes are pasted in a very random place probably related to cursor position.

I copy pasted much of the  selectAllNodes code and probably left some useless code

Improvements to be done: ability to paste inside a group

Don't hesitate to criticize: these are my very first line of code in C!