Fix OpenGL support handling for plugins with dynamic GL support.
OfxEffectInstance::onEnableOpenGLKnobValueChange() breaks rendering
for plugins that change their OpenGL support based on the value
of their parameters (e.g. OCIODisplay and friends). This method
can cause "OpenGL enabled" render calls to be made on plugins that
have explicitly stated they do not support OpenGL with their current
parameter values. This function also violates the contract between
host and plugin because it blindly changes the value of an instance
property without the plugin having any way to know it happened or
stop it. This logic is also not even needed because other logic in
Node::getCurrentOpenGLRenderSupport() actually properly handles all
the project, node, and plugin level overrides.

The following changes are include in this change:
- Removed OfxEffectInstance::onEnableOpenGLKnobValueChanged() and
  related declarations in the EffectInstace base class.
- Removed Project::isOpenGLRenderActivated() because it is
  basically identical to Project::isGPURenderingEnabledInProject().
- Clean up logic in callers that used to call
  onEnableOpenGLKnobValueChanged().
- Fixed a bug(?) where node "GPU render" knobs were not getting
  disabled when the project level knob caused GPU rendering to be
  disabled.
- Minor cleanups to make code simpler and more consistent