Make TrackerTest random numbers consistent across platforms.
Changed TrackerTests that use random numbers to use a specific
random number generator and seed so that the random numbers are
guaranteed to be the same no matter what platform you are on.
This ensures that we are testing with the same data on all
platforms.

The existing code used srand/rand which are not guaranteed
to be the same implementation across platforms and
it was leading to test failures on Windows that were not
occurring on Linux.