Remove threadHasACurrentContext() asserts
This change removes the asserts in OSGLContext::OSGLContext()
and OSGLContext::~OSGLContext() that check to see if the
context is set on the thread. These were invalid assumptions
because Qt appears to leave the current context set to its own
GL contexts when it dispatches event handlers for things like
button clicks. Removing the asserts avoid crashes and the
tests that were added in #871 make sure that Natron code
properly clears its own contexts.