Fix expansion-to-defined compiler warnings
This change fixes the tons of compiler warnings caused by using
"defined" in the COMPILER() and COMPILER_XX() macros. I simply
removed the defined expression since it isn't necessary. If the
preprocessor macro is not defined, it is given a value of 0 which
will return the same result as the (defined(X) && (X)) expressions
used in these macros.