Fix rendering bugs caused by intersecting rects with different mipMapLevels.
This change fixes 2 bugs where RectI objects containing information for
different mipMapLevels were being intersected with eachother. This was
causing visible compositing errors in the viewer when viewing the image
zoomed out and partially clipped by the edge of the viewer. This fix
just makes sure the RectIs are converted to the proper mipMapLevel before
being intersected.