Fix rendering bugs caused by intersecting rects with different mipMapLevels. (#918)
* Minor cleanup to help make it easier to verify correctness of fix.

- Updated rod computation so that rod could be declared as const.
  This is to make it easier to prove that this variable does not
  change beyond this point.

- Moved initial roi computation down to the image bounds computation
  code since that is where it is initially used and gets updated.

- Introduced isSupportedRenderScale() helper to remove duplicate code
  related to RenderScale support.

* Fix rendering bugs caused by intersecting rects with different mipMapLevels.

This change fixes 2 bugs where RectI objects containing information for
different mipMapLevels were being intersected with eachother. This was
causing visible compositing errors in the viewer when viewing the image
zoomed out and partially clipped by the edge of the viewer. This fix
just makes sure the RectIs are converted to the proper mipMapLevel before
being intersected.