Replace firebird2 dependency with firebird and refactor repo logic. (#939)
Msys2 has removed its firebird2 package. This change replaces the
package reference in our mingw-w64-natron-build-deps-qt5 package.

The pacman repo building workflow logic has been reworked so that
it generates versioned .zip files and builds the repo locally if
it can't download the current repo version.  This allows changes
to the repo to generate the new zip file AND allows the installer
and CI workflows to pass at the same time. This makes it easier
to verify that the changes to the repo actually fix build issues
when the pull request workflows run.