Fix crashing on startup for AMD GPUs.
This change makes sure there is an active context when
GetGPUInfoAMD() is called. While it is not explicitly stated in
https://www.opengl.org/registry/specs/AMD/wgl_gpu_association.txt
that a GL context needs to be current, it is somewhat implied by
the fact that these extension methods are fetched via
wglGetProcAddress(). According to the Windows docs, function
addresses are specific to a context so using them without an
active context seems to be undefined territory.