Change tools/utils/sourceList.py to always output posix paths.
Fixes the script so that it always outputs paths with forward
slashes even on Windows when using a Windows native Python build
instead of the msys2/cygwin one. Forward slash paths are needed
because this is the form qmake/cmake uses internally and is what
is expected when they call this build script.